% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions_for_DyMEP.R
\name{list_all_pmems_for_crop}
\alias{list_all_pmems_for_crop}
\title{A function to read list all available files for DyMEP}
\usage{
list_all_pmems_for_crop(crop_abbrev)
}
\arguments{
\item{crop_abbrev}{abbreviation for the  crop as defined in DyMEP package}
}
\description{
A function to read list all available files for DyMEP
}
\keyword{internal}
