% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions_for_DyMEP.R
\name{RDS_reader_pbCR}
\alias{RDS_reader_pbCR}
\title{A function to read DRC parameters files for DyMEP}
\usage{
RDS_reader_pbCR(crop_abbrev, file_name_best)
}
\arguments{
\item{crop_abbrev}{abbreviation for the  crop as defined in DyMEP package}

\item{file_name_best}{name of .rds DRC  for DyMEP package}
}
\description{
A function to read DRC parameters files for DyMEP
}
\keyword{internal}
