% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions_for_DyMEP.R
\name{RDS_reader_pmem}
\alias{RDS_reader_pmem}
\title{A function to read model files for DyMEP}
\usage{
RDS_reader_pmem(crop_abbrev, model_name)
}
\arguments{
\item{crop_abbrev}{abbreviation for the  crop as defined in DyMEP package}

\item{model_name}{name of .rds model for DyMEP package}
}
\description{
A function to read model files for DyMEP
}
\keyword{internal}
